ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
numpy==1.24.3
pandas==2.1.3
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
httpx==0.25.2

